Remove `metabase-types/types/Metadata`
