+++ frontend/src/metabase-lib/deprecated-types.ts
@@ -1,0 +1,31 @@
+import { Table } from "metabase-types/types/Table";
+import { Field } from "metabase-types/types/Field";
+
+type FieldsFilter = (fields: Field[]) => Field[];
+
+export type AggregationOperator = {
+  name: string;
+  short: string;
+  fields: Field[];
+  validFieldsFilters: FieldsFilter[];
+};
+
+type ValidArgumentsFilter = (field: Field, table: Table) => boolean;
+
+type FilterOperatorField = {
+  type: string;
+  values: {
+    name: string;
+    key: string;
+  }[];
+};
+
+export type FilterOperator = {
+  name: string;
+  verboseName: string;
+  moreVerboseName: string;
+  fields: FilterOperatorField[];
+  multi: boolean;
+  placeholders?: string[];
+  validArgumentsFilters: ValidArgumentsFilter[];
+};
