+++ frontend/src/metabase-lib/queries/structured/Aggregation.ts
@@ -2,13 +2,13 @@
 // @ts-nocheck
 import { t } from "ttag";
 import { Aggregation as AggregationObject } from "metabase-types/types/Query";
-import { AggregationOperator } from "metabase-types/types/Metadata";
 import { MetricId } from "metabase-types/types/Metric";
 import { FieldId } from "metabase-types/types/Field";
 import { TYPE } from "metabase-lib/types/constants";
 import * as AGGREGATION from "metabase-lib/queries/utils/aggregation";
 import Filter from "metabase-lib/queries/structured/Filter";
 import Metric from "metabase-lib/metadata/Metric";
+import type { AggregationOperator } from "metabase-lib/deprecated-types";
 import StructuredQuery from "../StructuredQuery";
 import Dimension, { AggregationDimension } from "../../Dimension";
 import MBQLClause from "./MBQLClause";
